BatchTime/WebappDemoplatform/requirements.txt
# azure-core==1.24.2
azure-storage-blob==12.13.1
# certifi==2021.5.30
# cffi==1.15.1
# charset-normalizer==2.0.12
# click==8.0.4
# colorama==0.4.5
# cryptography==40.0.1
# dataclasses==0.8
Flask==2.0.3
# idna==3.4
# importlib-metadata==4.8.3
# isodate==0.6.1
# itsdangerous==2.0.1
# Jinja2==3.0.3
# MarkupSafe==2.0.1
# msrest==0.7.1
# numpy==1.19.5
# oauthlib==3.2.2
pandas==1.1.5
# pycparser==2.21
# python-dateutil==2.8.2
# pytz==2023.2
# requests==2.27.1
# requests-oauthlib==1.3.1
# six==1.16.0
# typing_extensions==4.1.1
# urllib3==1.26.15
Werkzeug==2.0.3
# wincertstore==0.2
# zipp==3.6.0
